updated to current registration feature on backend

+++ controllers/userCtrls.js
@@ -1,8 +1,17 @@
 const express = require('express')
 const bcrypt = require('bcrypt')
 const db = require('../models')
+const { query } = require('express')
 console.log(db)
 const router = express.Router()
+
+// AUTH REQUIRED MIDDLEWARE //
+const authRequired = (req, res, next) => {
+    if (!req.session.currentUser) {
+      return res.redirect('/signin')
+    }
+    next()
+}
 
 // FIND USER //
 const getUser = (req, res) => {
@@ -17,19 +26,32 @@ const getUser = (req, res) => {
 }
 
 // CREATE USER //
-const createUser = (req, res) => {
-    db.Users.create(req.body)
-    .then((createdUser) => {
-        if(!createdUser) {
-            res.status(404).json({message: 'Cannot create User'})
-        } else {
-            res.status(201).json({data: createdUser})
-        }
-    })
+const createUser = async (req, res) => {
+    console.log(req.body)
+    try {
+        const salt = bcrypt.genSaltSync(10);
+        const encryptedPassword = bcrypt.hashSync(req.body.password, salt);
+        
+        const createdUser = await db.Users.create({
+        name: req.body.name,
+        password: encryptedPassword
+    });
+
+    if (!createdUser) {
+        res.status(404).json({message: 'Cannot create User'});
+    } else {
+        res.status(201).json({data: createdUser})
+    }
+    } catch (err) {
+        console.log(err)
+    }
 }
 
 // DELETE USER //
 const deleteUser = (req, res) => {
+    if (!req.session.currentUser) {
+        return res.status(401).json({ message: 'Unauthorized' })
+    }
     db.Users.findByIdAndDelete(req.params.id)
     .then((deletedUser) => {
         if(!deletedUser) {
@@ -42,85 +64,73 @@ const deleteUser = (req, res) => {
 
 /// TESTING SIGN IN STUFF ///
 /////////////////////////////
-// SIGN UP PAGE //
-// router.post('/signup', (req, res) => {
-//     // encrypt pws with salt
-//     const salt = bcrypt.genSaltSync(10);
-//     console.log(req.body); // password
-//     req.body.password = bcrypt.hashSync(req.body.password, salt); // encrypt 
-//     console.log(req.body); // encrypted pw
-
-//     // Check if username already exists
-//     User.findOne({ name: req.body.name })
-//         .then(userExists => {
-//             if(userExists) {
-//                 res.send('That username is taken');
-//             } else {
-//                 User.create(req.body)
-//                     .then(createdUser => {
-//                         console.log(createdUser);
-//                         req.session.currentUser = createdUser;
-//                         res.redirect('/scores');
-//                     })
-//                 .catch(err => console.log(err));
-//             }
-//         })
-//         .catch(err => console.log(err));
-// });
-
-// SIGN IN PAGE
-// router.get('/signin', (req, res) => {
-//     res.send('this is the signin page');
-// });
-
-// SIGN UP PAGE //
-// router.post('/signup', (req, res) => {
-//     // encrypt pws with salt
-//     const salt = bcrypt.genSaltSync(10);
-//     console.log(req.body); // password
-//     req.body.password = bcrypt.hashSync(req.body.password, salt); // encrypt 
-//     console.log(req.body); // encrypted pw
-
-//     // Check if username already exists
-//     User.findOne({ name: req.body.name })
-//         .then(userExists => {
-//             if(userExists) {
-//                 res.send('That username is taken');
-//             } else {
-//                 User.create(req.body)
-//                     .then(createdUser => {
-//                         console.log(createdUser);
-//                         req.session.currentUser = createdUser;
-//                         res.redirect('/scores');
-//                     })
-//                 .catch(err => console.log(err));
-//             }
-//         })
-//         .catch(err => console.log(err));
-// })
+// SIGN UP PAGE (Register) //
+// const signUp = async (req, res) => {
+//     try {
+//       // encrypt pws with salt
+//       const salt = bcrypt.genSaltSync(10);
+//       console.log(req.body); // password
+//       req.body.password = bcrypt.hashSync(req.body.password, salt); // encrypt 
+//       console.log(req.body); // encrypted pw
+  
+//       // Check if username already exists
+//       const userExists = await User.findOne({ name: req.body.name });
+//       if (userExists) {
+//         res.send('That username is taken');
+//       } else {
+//         const createdUser = await User.create(req.body);
+//         console.log(createdUser);
+//         req.session.currentUser = createdUser;
+//         res.redirect('/scorecast');
+//       }
+//     } catch (err) {
+//       console.log(err);
+//     }
+// };
+  
 
 // SIGN IN FUNCTION //
-// router.post('/signin', (req, res) => {
-//     // find user
-//     User.findOne({ username: req.body.username })
+// const signIn = (req, res) => {
+//     const { username, password } = req.body;
+  
+//     User.findOne({ username })
 //       .then((foundUser) => {
 //         if (foundUser) {
-//           // compare passwords w/ bcrpt
-//           const validLogin = bcrypt.compareSync(req.body.password, foundUser.password);
-//           // compareSync returns true or false
+//           const validLogin = bcrypt.compareSync(password, foundUser.password);
+  
 //           if (validLogin) {
 //             req.session.currentUser = foundUser;
-//             // session is logged in
-//             res.redirect('/scores');
-//             } else {
+//             res.redirect('/scorecast');
+//           } else {
 //             res.send('Invalid username or password');
-//             }
+//           }
+//         } else {
+//           res.send('Invalid username or password');
 //         }
 //     })
-// })
+//     .catch((err) => console.log(err));
+// };  
+
+////// SEARCH BAR TEST //////
+// SEARCH BAR METHOD/ROUTE //
+// const searchScore = (req, res) => {
+//     const search = req.query.q
+//     db.Users.find({ name: {$regex: query, $options: 'i' } })
+//     .then((foundScore) => {
+//         if (!foundScore) {
+//             res.status(404).json({message: `No ${query} found`})
+//         } else {
+//             res.status(200).json({data: foundScore})
+//         }
+//     })
+//     .catch((err) => {
+//         res.status(500).json({ message: 'Server error', error: err})
+//     })
+// }
 
 module.exports = {
     createUser,
     deleteUser,
-    getUser
+    getUser,
+    authRequired
 }
